LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/shape-functions/ellipse-function-valid-expected.txt


PASS e.style['shape-outside'] = "ellipse()" should set the property value
PASS e.style['shape-outside'] = "ellipse(1px 2px)" should set the property value
FAIL e.style['shape-outside'] = "ellipse(20px 40px at center)" should set the property value assert_equals: serialization should be canonical expected "ellipse(20px 40px at 50% 50%)" but got "ellipse(20px 40px at center center)"
PASS e.style['shape-outside'] = "ellipse(closest-side 20%)" should set the property value
PASS e.style['shape-outside'] = "ellipse(farthest-side 20%)" should set the property value
PASS e.style['shape-outside'] = "ellipse(closest-corner 20%)" should set the property value
PASS e.style['shape-outside'] = "ellipse(farthest-corner 20%)" should set the property value
PASS e.style['shape-outside'] = "ellipse(at 10% 20%)" should set the property value
PASS e.style['shape-outside'] = "ellipse(at -10px -20%)" should set the property value
FAIL e.style['shape-outside'] = "ellipse(4% 20% at top right)" should set the property value assert_equals: serialization should be canonical expected "ellipse(4% 20% at 100% 0%)" but got "ellipse(4% 20% at right top)"
PASS e.style['shape-outside'] = "ellipse(calc(100% - 20px) calc(80% - 10px) at calc(100% - 20px) calc(100% / 4))" should set the property value
FAIL e.style['shape-outside'] = "ellipse(10px closest-side at top right)" should set the property value assert_equals: serialization should be canonical expected "ellipse(10px closest-side at 100% 0%)" but got "ellipse(10px closest-side at right top)"
FAIL e.style['shape-outside'] = "ellipse(farthest-side 20px at center top)" should set the property value assert_equals: serialization should be canonical expected "ellipse(farthest-side 20px at 50% 0%)" but got "ellipse(farthest-side 20px at center top)"
FAIL e.style['shape-outside'] = "ellipse(farthest-side farthest-side at top right)" should set the property value assert_equals: serialization should be canonical expected "ellipse(farthest-side farthest-side at 100% 0%)" but got "ellipse(farthest-side farthest-side at right top)"
